reorganize TreeSummary for multiple traces
